# internal/api/rest/gen.go
// Package rest provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package rest

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for GetPostsParamsSort.
const (
	PostedAtasc  GetPostsParamsSort = "postedAt,asc"
	PostedAtdesc GetPostsParamsSort = "postedAt,desc"
	Titleasc     GetPostsParamsSort = "title,asc"
	Titledesc    GetPostsParamsSort = "title,desc"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code The resulting http code
	Code    int         `json:"code"`
	Details interface{} `json:"details"`

	// Message Description of the error occurred
	Message string `json:"message"`
}

// PageMetadata defines model for PageMetadata.
type PageMetadata struct {
	// Page The returned page
	Page int `json:"page"`

	// Size The size of the page
	Size int `json:"size"`

	// TotalCount The total number of items
	TotalCount int `json:"totalCount"`

	// TotalPages The total number of pages
	TotalPages int `json:"totalPages"`
}

// PagePostComments defines model for PagePostComments.
type PagePostComments struct {
	UnderscoreMetadata PageMetadata  `json:"_metadata"`
	Data               []PostComment `json:"data"`
}

// PagePosts defines model for PagePosts.
type PagePosts struct {
	UnderscoreMetadata PageMetadata `json:"_metadata"`
	Data               []Post       `json:"data"`
}

// Post defines model for Post.
type Post struct {
	// Content The post content
	Content string `json:"content"`

	// CreatedAt Creating time of the comment
	CreatedAt time.Time `json:"createdAt"`

	// ID Id of the post
	ID uint `json:"id"`

	// PostedAt Posting time of the comment
	PostedAt time.Time `json:"postedAt"`

	// Self Self link of the post
	Self string `json:"self"`

	// Title Title of the post
	Title string `json:"title"`

	// UpdatedAt Updating time of the comment
	UpdatedAt time.Time `json:"updatedAt"`

	// UserID Id of the user
	UserID uint `json:"userId"`
}

// PostComment defines model for PostComment.
type PostComment struct {
	// Comment The comment value
	Comment string `json:"comment"`

	// CreatedAt Creating time of the comment
	CreatedAt time.Time `json:"createdAt"`

	// ID Id of the comment
	ID uint `json:"id"`

	// PostID Id of the post
	PostID uint `json:"postId"`

	// PostedAt Posting time of the comment
	PostedAt time.Time `json:"postedAt"`

	// UpdatedAt Updating time of the comment
	UpdatedAt time.Time `json:"updatedAt"`

	// UserID Id of the user who wrote the comment
	UserID uint `json:"userId"`
}

// PostCommentNew defines model for PostCommentNew.
type PostCommentNew struct {
	// Comment The comment value
	Comment string `json:"comment"`

	// Csrf CSRF token
	Csrf openapi_types.UUID `json:"csrf"`

	// UserID Id of the user who wrote the comment
	UserID uint `json:"userId"`
}

// User defines model for User.
type User struct {
	// CreatedAt Creating time of the user
	CreatedAt time.Time `json:"createdAt"`

	// ID Id of the user
	ID uint `json:"id"`

	// IsAdmin Whether the user is admin or not
	IsAdmin bool `json:"isAdmin"`

	// Password Password of the user
	Password string `json:"password"`

	// Self Self link of the user
	Self string `json:"self"`

	// UpdatedAt Updating time of the user
	UpdatedAt time.Time `json:"updatedAt"`

	// Username Username of the user
	Username string `json:"username"`
}

// UserCredential defines model for UserCredential.
type UserCredential struct {
	IsAdmin *bool `json:"isAdmin,omitempty"`

	// Password Password of the user
	Password string `json:"password"`

	// Username Username of the user
	Username string `json:"username"`
}

// GetPostsParams defines parameters for GetPosts.
type GetPostsParams struct {
	// Page Page to retrieve
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// Size Size of the page
	Size *int `form:"size,omitempty" json:"size,omitempty"`

	// Sort Sorting criteria
	Sort *GetPostsParamsSort `form:"sort,omitempty" json:"sort,omitempty"`
}

// GetPostsParamsSort defines parameters for GetPosts.
type GetPostsParamsSort string

// GetPostCommentsParams defines parameters for GetPostComments.
type GetPostCommentsParams struct {
	// Page Page to retrieve
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// Size Size of the page
	Size *int `form:"size,omitempty" json:"size,omitempty"`
}

// PostPostCommentJSONRequestBody defines body for PostPostComment for application/json ContentType.
type PostPostCommentJSONRequestBody = PostCommentNew

// UserLoginJSONRequestBody defines body for UserLogin for application/json ContentType.
type UserLoginJSONRequestBody = UserCredential

// UserSignupJSONRequestBody defines body for UserSignup for application/json ContentType.
type UserSignupJSONRequestBody = UserCredential

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// get posts
	// (GET /api/posts)
	GetPosts(c *gin.Context, params GetPostsParams)
	// get post
	// (GET /api/posts/{postId})
	GetPostById(c *gin.Context, postID uint)
	// get comments page
	// (GET /api/posts/{postId}/comments)
	GetPostComments(c *gin.Context, postID uint, params GetPostCommentsParams)
	// post comment
	// (POST /api/posts/{postId}/comments)
	PostPostComment(c *gin.Context, postID uint)
	// Login
	// (POST /api/users/login)
	UserLogin(c *gin.Context)
	// Logout
	// (DELETE /api/users/logout)
	UserLogout(c *gin.Context)
	// User Signup endpoint
	// (POST /api/users/signup)
	UserSignup(c *gin.Context)
	// get user
	// (GET /api/users/{userId})
	GetUserById(c *gin.Context, userID uint)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetPosts operation middleware
func (siw *ServerInterfaceWrapper) GetPosts(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetPostsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "size" -------------

	err = runtime.BindQueryParameter("form", true, false, "size", c.Request.URL.Query(), &params.Size)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter size: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "sort" -------------

	err = runtime.BindQueryParameter("form", true, false, "sort", c.Request.URL.Query(), &params.Sort)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sort: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPosts(c, params)
}

// GetPostById operation middleware
func (siw *ServerInterfaceWrapper) GetPostById(c *gin.Context) {

	var err error

	// ------------- Path parameter "postId" -------------
	var postID uint

	err = runtime.BindStyledParameterWithOptions("simple", "postId", c.Param("postId"), &postID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter postId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPostById(c, postID)
}

// GetPostComments operation middleware
func (siw *ServerInterfaceWrapper) GetPostComments(c *gin.Context) {

	var err error

	// ------------- Path parameter "postId" -------------
	var postID uint

	err = runtime.BindStyledParameterWithOptions("simple", "postId", c.Param("postId"), &postID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter postId: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetPostCommentsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "size" -------------

	err = runtime.BindQueryParameter("form", true, false, "size", c.Request.URL.Query(), &params.Size)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter size: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPostComments(c, postID, params)
}

// PostPostComment operation middleware
func (siw *ServerInterfaceWrapper) PostPostComment(c *gin.Context) {

	var err error

	// ------------- Path parameter "postId" -------------
	var postID uint

	err = runtime.BindStyledParameterWithOptions("simple", "postId", c.Param("postId"), &postID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter postId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostPostComment(c, postID)
}

// UserLogin operation middleware
func (siw *ServerInterfaceWrapper) UserLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UserLogin(c)
}

// UserLogout operation middleware
func (siw *ServerInterfaceWrapper) UserLogout(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UserLogout(c)
}

// UserSignup operation middleware
func (siw *ServerInterfaceWrapper) UserSignup(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UserSignup(c)
}

// GetUserById operation middleware
func (siw *ServerInterfaceWrapper) GetUserById(c *gin.Context) {

	var err error

	// ------------- Path parameter "userId" -------------
	var userID uint

	err = runtime.BindStyledParameterWithOptions("simple", "userId", c.Param("userId"), &userID, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter userId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetUserById(c, userID)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/posts", wrapper.GetPosts)
	router.GET(options.BaseURL+"/api/posts/:postId", wrapper.GetPostById)
	router.GET(options.BaseURL+"/api/posts/:postId/comments", wrapper.GetPostComments)
	router.POST(options.BaseURL+"/api/posts/:postId/comments", wrapper.PostPostComment)
	router.POST(options.BaseURL+"/api/users/login", wrapper.UserLogin)
	router.DELETE(options.BaseURL+"/api/users/logout", wrapper.UserLogout)
	router.POST(options.BaseURL+"/api/users/signup", wrapper.UserSignup)
	router.GET(options.BaseURL+"/api/users/:userId", wrapper.GetUserById)
}